refactor: move  audiobattle to components folder
