+++ src/components/AudioBattle/AudioBatle.css
@@ -174,7 +174,7 @@
 }
 
 .audio-battle__repeat-audio-list {
-  background-image: url(../../../assets/icons/audio.svg);
+  background-image: url(../../assets/icons/audio.svg);
   background-position: center;
   background-color: aliceblue;
   border-radius: 50%;
@@ -196,7 +196,7 @@
 }
 
 .audio-battle__repeat-audio_small {
-  background-image: url(../../../assets/icons/audio.svg);
+  background-image: url(../../assets/icons/audio.svg);
   background-position: center;
   background-color: aliceblue;
   border-radius: 50%;
@@ -218,7 +218,7 @@
 }
 
 .audio-battle__repeat-audio_big {
-  background-image: url(../../../assets/icons/audio.svg);
+  background-image: url(../../assets/icons/audio.svg);
   background-position: center;
   background-color: aliceblue;
   border-radius: 50%;
