+++ src/components/AudioBattle/AudioBatle.scss
@@ -22,7 +22,7 @@
 @mixin commonRepeatAudio($bgSize, $width, $heigt) {
   // background-image: url(../../../assets/speaker.png);
 
-  background-image: url(../../../assets/icons/audio.svg);
+  background-image: url(../../assets/icons/audio.svg);
   background-position: center;
   background-color: aliceblue;
   border-radius: 50%;
