+++ src/components/Common/EndGameStatistic/StatTable.tsx
